% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workspace.R
\name{clean_workspace}
\alias{clean_workspace}
\title{Clean your workspace}
\usage{
clean_workspace(force = FALSE, silent = justifier::opts$get("silent"))
}
\arguments{
\item{force}{Whether to force cleaning the workspace}

\item{silent}{Whether to be chatty or silent.}
}
\description{
Clean your workspace
}
\examples{
### Without `force=TRUE`, presents a query to the user in
### interactive mode:
clean_workspace(silent=FALSE);

### Set `force=TRUE` to force clean the workspace
clean_workspace(force = TRUE, silent=FALSE);
}
