% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workspace.R
\name{get_workspace}
\alias{get_workspace}
\title{Get your justifier workspace identifier}
\usage{
get_workspace(silent = justifier::opts$get("silent"))
}
\arguments{
\item{silent}{Whether to be suppress messages.}
}
\value{
Invisibly, the workspace identifier.
}
\description{
This is used to be able to log decisions programmatically.
}
\examples{
justifier::get_workspace_id();
}
