% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_specLists.R
\name{merge_specLists}
\alias{merge_specLists}
\title{Merging to justifier specification lists}
\usage{
merge_specLists(x, y)
}
\arguments{
\item{x, y}{The two justifier specification lists}
}
\value{
A merged justifier specification list.
}
\description{
Merging to justifier specification lists
}
\examples{
### Add example
}
