% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_justifications.R
\name{parse_justifications}
\alias{parse_justifications}
\title{Parsing justifications}
\usage{
parse_justifications(x)
}
\arguments{
\item{x}{The `justifier` object that has just been loaded by `load_justifications`.}
}
\value{
The parsed `justifier` object
}
\description{
While there is some flexibility in how justifications can be specified,
they are most easily processed further if they all follow the same
conventions. This function ensures this. The convention is as follows:
- all specifications are provided in four 'flat' lists, named after the
  types of elements they contain;
- all elements have  a unique identifier
- all references to other elements are indeed only references to the other
  elements' id's in these 'flat lists'
}
