% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_workspace_id.R
\name{set_workspace_id}
\alias{set_workspace_id}
\title{Set your justifier workspace identifier}
\usage{
set_workspace_id(id, silent = justifier::opts$get("silent"))
}
\arguments{
\item{id}{The workspace identifier}

\item{silent}{Whether to be suppress messages.}
}
\value{
Invisibly, the passed \code{id}.
}
\description{
This is used to be able to log decisions programmatically.
}
\examples{
set_workspace_id("my_workspace");
}
