% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.justifierStructured.R
\name{c.justifierStructured}
\alias{c.justifierStructured}
\title{Concatenate two or more structured justifier objects}
\usage{
\method{c}{justifierStructured}(...)
}
\arguments{
\item{...}{Structured justifier objects}
}
\value{
Invisibly, the concatenated list
}
\description{
Concatenate two or more structured justifier objects
}
