% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace}
\alias{workspace}
\title{Show your workspace contents}
\usage{
workspace(silent = justifier::opts$get("silent"))
}
\arguments{
\item{silent}{Whether to be chatty or silent.}
}
\value{
The workspace contents.
}
\description{
Show your workspace contents
}
\examples{
justifier::clean_workspace(force = TRUE, silent=FALSE);
justifier::log_decision(
  "First we start using `justifier`.",
  silent=FALSE
);
justifier::log_decision(
  paste0("Then we start documenting our ",
         "decisions and justifications."),
  silent=FALSE
);
justifier::log_decision(
  "Then we start learning from ourselves.",
  silent=FALSE
);
justifier::workspace();
}
