% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srce.R
\name{source}
\alias{source}
\alias{dcsn}
\alias{jstf}
\alias{asrt}
\alias{srce}
\alias{assert}
\alias{justify}
\alias{c.justifierElement}
\title{Programmatically constructing justifier elements}
\usage{
source(label, description = NULL, xdoi = NULL, type = NULL, id = NULL, ...)

assert(label, description = "", id = NULL, source = NULL, ...)

justify(label, description = "", id = NULL, assertion = NULL, ...)

dcsn(label, description = NULL, id = NULL, justification = NULL, ...)

\method{c}{justifierElement}(...)
}
\arguments{
\item{label}{A human-readable label for the \code{decision}, \code{justification},
\code{assertion}, or \code{source}.}

\item{description}{A human-readable description.}

\item{xdoi, type}{For \code{source}s, XDOI identifier and type can also be
specified.}

\item{id}{The identifier (randomly generated if omitted).}

\item{...}{Additional fields and values to store in the element.}

\item{source}{In assertions, the source (or sources) that the assertion
is based on can be specified using \code{srce()}.}

\item{assertion}{In justifications, the assertion (or assertions) that
the justification is based on can be specified using \code{asrt()}.}

\item{justification}{In decisions, the justification (or justifications)
that the decision is based on can be specified using \code{jstf()}.}
}
\value{
The generated object.
}
\description{
These functions can be used to programmatically construct justifications.
}
\examples{
### Programmatically create a simple justification object
justifierObject <-
  justifier::asrt(
    "assertion",
    source = c(
      justifier::srce('source1'),
      justifier::srce('source2')));
}
