% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srce.R
\name{srce}
\alias{srce}
\alias{asrt}
\alias{jstf}
\alias{dcsn}
\alias{c.justifierElement}
\title{Programmatically constructing justifier elements}
\usage{
srce(label, description = NULL, xdoi = NULL, type = NULL, id = NULL, ...)

asrt(label, description = NULL, id = NULL, source = NULL, ...)

jstf(label, description = NULL, id = NULL, assertion = NULL, ...)

dcsn(label, description = NULL, id = NULL, justification = NULL, ...)

\method{c}{justifierElement}(...)
}
\arguments{
\item{label}{A human-readable label for the \code{decision}, \code{justification},
\code{assertion}, or \code{source}.}

\item{description}{A human-readable description.}

\item{xdoi, type}{For \code{source}s, XDOI identifier and type can also be
specified.}

\item{id}{The identifier (randomly generated if omitted).}

\item{...}{Additional fields and values to store in the element.}
}
\value{
The generated object.
}
\description{
These functions can be used to programmatically construct justifications.
}
\examples{
### Programmatically create a simple justification object
justifierObject <-
  justifier::asrt(
    "assertion",
    source = c(
      justifier::srce('source1'),
      justifier::srce('source2')));
}
