% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.justifierStructuredObject.R
\name{c.justifierStructuredObject}
\alias{c.justifierStructuredObject}
\title{Concatenate two or more structured justifier objects}
\usage{
\method{c}{justifierStructuredObject}(...)
}
\arguments{
\item{...}{Structured justifier objects}
}
\value{
Invisibly, the concatenated list
}
\description{
Concatenate two or more structured justifier objects
}
